requirements.txt
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-huggingface
faiss-cpu
sentence-transformers
gitpython
rank-bm25
google-generativeai
protobuf==4.25.3
requests
typing_extensions
python-dotenv
numpy
dotenv
fastapi
uvicorn[standard]
gunicorn
pydantic>=2.0